add global config file
In the config file some default settings, like which broker to use, can be selected.
This file is made as a JS file so that the website can load in this file on startup.
The client needs to read this file and remove the first and last line. After this,
the file can be used in any JSON parser.